Multiple deprecation-related issues addressed:
DATA_RATING is deprecated and was not marked as such. Fixes Issue #1.

llRemoteDataSetRegion is deprecated and was not marked as such;
llCloud was deprecated as of Second Life Server 13.02.08.270166.
These two changes together fix Issue #2. Thanks to makonozakisl
for the report.

AGENT is no longer deprecated as per BUG-2830. Its text and the text
for AGENT_BY_LEGACY_NAME and AGENT_BY_USERNAME have been tweaked to
reflect the situation.

Descriptions for llSound, llSoundPreload and llMake* now indicate
explicitly that the function is deprecated.