firestorm_lscript_library.cpp.in was unused due to a bug in generate.sh which was using the wrong file.
As a result, a bug in the source output module was not caught for a
long time, resulting in firestorm_lscript_library.cpp output file
not being updated at all. The Firestorm project decided not to use it
anyway, so that output file doesn't matter and will be removed soon.