Make checking for string return value stricter in functions check.
LSL adds implicit casts from key to string in most situations when the original value is a key. Therefore, assigning the return value to a string variable doesn't suffice to ensure that the type is correct. However, since it's not valid to concatenate a string and a key, that can be used to raise an alarm if the original function didn't actually return string.

I haven't figured out how to do the equivalent with floats. Assigning a float value to an integer will give a type mismatch, but I'm not aware of any context in which an integer value is not valid when a float is expected.

Similarly with keys: I'm not aware of any context in which a string value is not valid when a key is expected.

Unfortunately, only return values can be checked this way. Parameters can't.