Add SIT_* constants; add constant-test-*.lsl
Commit 101:c50ed747e4d1 introduced the force sit functions and constants, but it neglected to introduce the seven constants that represent the return status of llSitOnLink.

We introduce them here, and we also add to the repository the pre-generated constant-test-(sl|os|aa).lsl, which weren't under version control.