Accept backslash-escaped strings
The XML parser rejects most control characters including the newly introduced NAK character (U+0015), so we now allow C-style backslash-escaped strings in the XML attributes (in particular for the values of constants), which are now unescaped before further processing. It's up to the output modules to perform any additional escaping.